+++ compiler-plugin/src/main/kotlin/tech/mappie/ir/resolving/enums/EnumResolver.kt
@@ -1,10 +1,13 @@
 package tech.mappie.ir.resolving.enums
 
 import org.jetbrains.kotlin.ir.declarations.IrFunction
+import org.jetbrains.kotlin.ir.expressions.IrFunctionExpression
 import org.jetbrains.kotlin.ir.types.IrType
 import org.jetbrains.kotlin.ir.types.getClass
+import org.jetbrains.kotlin.ir.util.statements
 import tech.mappie.ir.resolving.MappingResolver
 import tech.mappie.ir.resolving.ResolverContext
+import tech.mappie.ir.resolving.findMappingStatements
 
 class EnumResolver(
     private val context: ResolverContext,
@@ -16,8 +19,14 @@ class EnumResolver(
         EnumMappingRequestBuilder(source, target)
             .sources(source.getClass()!!.accept(EnumEntriesCollector(), Unit))
             .targets(target.getClass()!!.accept(EnumEntriesCollector(), Unit))
-            .also { function?.body?.accept(EnumMappingBodyCollector(), it) }
-            .construct(context.origin!!)
+            .apply {
+                val mapping = findMappingStatements(function?.body).singleOrNull()?.arguments?.getOrNull(1) as? IrFunctionExpression
+                mapping?.function?.body?.statements?.forEach { statement ->
+                    statement.accept(EnumMappingStatementCollector, Unit)
+                        ?.let { explicit(it) }
+                }
+            }
+            .construct(context.origin)
             .let { listOf(it) }
 }