+++ website/src/changelog.md
@@ -5,7 +5,10 @@ changelog:
   - date: "tbd"
     title: "v2.2.20-1.6.0"
     items:
+      - "Allow `mapping` to be used anywhere (at most once) in the body of the map function."
       - "Fixed check of runtime usage of of mapping DSL for multiple object mappie."
+      - "Minor performance enhancements - by @ErwinOlie."
+      - "Fix for missing reporting when both a mapper and its generated mapper have issues - by @ErwinOlie."
   - date: "2025-09-10"
     title: "v2.2.20-1.5.0"
     items: